controllers working with get requests

+++ backend/HospitalManagement/src/main/java/com/cg/dto/UndergoesDTO.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 
 public class UndergoesDTO {
 
-	private Integer patientId;
+	private Long patientId;
 
 	private Integer procedureCode;
 
@@ -21,7 +21,7 @@ public class UndergoesDTO {
 	}
 
 	// Parameterized constructor
-	public UndergoesDTO(Integer patientId, Integer procedureCode, Integer stayId, LocalDateTime dateUndergoes,
+	public UndergoesDTO(Long patientId, Integer procedureCode, Integer stayId, LocalDateTime dateUndergoes,
 			Integer physicianId, Integer assistingNurseId) {
 		this.patientId = patientId;
 		this.procedureCode = procedureCode;
@@ -32,11 +32,11 @@ public class UndergoesDTO {
 	}
 
 	// Getters and Setters
-	public Integer getPatientId() {
+	public Long getPatientId() {
 		return patientId;
 	}
 
-	public void setPatientId(Integer patientId) {
+	public void setPatientId(Long patientId) {
 		this.patientId = patientId;
 	}
